update tests for tool, add new tests for click
Because the tool is using the Person.csv from a new branch (not master), plus found Monet exists in ReadAct so that Musk is used to replace Monet as the sample of new person entry, therefore, the tests for the command-line tool are updated.

Add new tests for click, to be continued.

see #56 #41 #37 #32 #63